Reworking the CODES infrastructure
Each type of code is now a Global type or it has a specific Region (US, State, etc.).
This allows us to have different code values for different departments in different
regions (KY, Canada, etc.)

I also worked in some of the new code tables, which are more complete than my first
set of data.  The new ones are in UPPER CASE but I figure we can fix that later if
we need to.  Or we can just use CSS to display it to the user properly.  :)