Added a height to the sidebar logo area.
Refreshing the page was causing the sidebar buttons to jump around.  So, I figured
that telling the logo how tall it was suposed to be would keep that from happening.
It looks like it worked, but I only tested it on Linux/Chrome so there could be all
kinds of problems with it.  :)